@@ -10,29 +10,25 @@
   "scripts": {
     "build": "npx webpack --config ./src/webpack.config.js --mode production",
     "dev": "npx webpack --config ./src/webpack.config.js --mode development --watch",
-    "test": "npx standard && npx jest"
+    "test": "npx eslint src/ && npx vitest run"
   },
   "author": "githubtraining",
   "license": "MIT",
-  "standard": {
-    "env": [
-      "jest",
-      "browser"
-    ]
-  },
   "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.2.3",
-    "@babel/register": "^7.9.0",
-    "babel-jest": "^29.7.0",
-    "babel-loader": "^8.0.5",
+    "@babel/core": "^7.24.0",
+    "@babel/preset-env": "^7.24.0",
+    "@babel/register": "^7.24.0",
+    "babel-loader": "^9.2.0",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
     "html-webpack-inject-attributes-plugin": "^1.0.6",
-    "html-webpack-plugin": "^5.5.0",
+    "html-webpack-plugin": "^5.6.0",
     "html-webpack-tags-plugin": "^3.0.2",
-    "standard": "^14.3.4",
-    "webpack": "^5.89.0",
+    "prettier": "^3.1.0",
+    "vitest": "^1.0.0",
+    "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "webpack-dev-server": "^5.0.0"
   },
   "dependencies": {
     "node-forge": "^1.3.0"
